+++ src/Rasa.Game/Packets/MapChannel/Server/TeleportPacket.cs
@@ -10,12 +10,12 @@ namespace Rasa.Packets.MapChannel.Client
         public override GameOpcode Opcode { get; } = GameOpcode.Teleport;
 
         public Vector3 Position { get; set; }           // position
-        public float Yaw { get; set; }                  // yaw
+        public double Yaw { get; set; }                  // yaw
         public TeleportType TeleportType { get; set; }  // teleportType
         public uint Delay { get; set; }                 // delay
         public bool DoCancel { get; set; }              // doCancel
 
-        public TeleportPacket(Vector3 position, float yaw, TeleportType teleportType, uint delay)
+        public TeleportPacket(Vector3 position, double yaw, TeleportType teleportType, uint delay)
         {
             Position = position;
             Yaw = yaw;
